rename Logger to GlobalLogger, BaseLogger to Logger. Added new registerContextAwareLoggerToContainer function.
