+++ packages/logger/src/base-console-logger-client.ts
@@ -6,7 +6,7 @@ import * as _ from 'lodash';
 import * as utils from 'util';
 
 import { AvailabilityTelemetry } from './availablity-telemetry';
-import { LogLevel } from './base-logger';
+import { LogLevel } from './logger';
 import { BaseTelemetryProperties } from './base-telemetry-properties';
 import { LoggerClient } from './logger-client';
 import { LoggerEvent } from './logger-event';
