+++ packages/logger/src/index.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-export { Logger } from './logger';
+export { GlobalLogger } from './global-logger';
 export { ContextAwareLogger } from './context-aware-logger';
-export { LogLevel } from './base-logger';
+export { LogLevel } from './logger';
 export { loggerTypes } from './logger-types';
 export { registerLoggerToContainer } from './register-logger-to-container';
 export { BaseTelemetryProperties } from './base-telemetry-properties';
@@ -18,3 +18,4 @@ export {
 } from './logger-event-measurements';
 export { AvailabilityTelemetry } from './availablity-telemetry';
 export { LoggerProperties } from './logger-properties';
+export { Logger } from './logger';
