+++ packages/logger/src/logger-client.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { AvailabilityTelemetry } from './availablity-telemetry';
-import { LogLevel } from './base-logger';
+import { LogLevel } from './logger';
 import { LoggerEvent } from './logger-event';
 import { TelemetryMeasurements } from './logger-event-measurements';
 import { LoggerProperties } from './logger-properties';
